server/requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
easyocr==1.6.2
Flask==2.2.2
Flask-Cors==3.0.10
idna==3.4
imageio==2.22.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
networkx==2.8.8
ninja==1.11.1
numpy==1.23.5
opencv-python-headless==4.5.4.60
packaging==22.0
Pillow==9.3.0
pyclipper==1.3.0.post4
python-bidi==0.4.2
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.3
Shapely==1.8.5.post1
six==1.16.0
tifffile==2022.10.10
torch==1.13.0
torchvision==0.14.0
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
